[PCFG-3] fix swagger
